requirements.txt
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.32.0
django-rest-auth==0.9.1
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
oauthlib==2.0.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
